requirements.txt
loguru==0.7.2
paramiko==3.5.0
python-telegram-bot==13.7
python-dotenv==1.0.1
